remove most of the warning
[ROCm/MIOpen commit: d7b79820974bff472e53f32a1034e2de0652b9f7]

+++ src/tensor.cpp
@@ -47,7 +47,7 @@ const std::vector<int>& TensorDescriptor::GetStrides() const
 int TensorDescriptor::GetSize() const
 {
 	assert(lens.size() == strides.size());
-	return lens.size();
+	return (int)lens.size();
 }
 int TensorDescriptor::GetElementSize() const
 {
